refactor: split some convert function into `sql-common` crates
