+++ src/sql/Cargo.toml
@@ -19,6 +19,7 @@ common-decimal.workspace = true
 common-error.workspace = true
 common-macro.workspace = true
 common-query.workspace = true
+common-sql.workspace = true
 common-time.workspace = true
 datafusion.workspace = true
 datafusion-common.workspace = true
